move arithmetic impls into submodule
