+++ ogma/src/lang/impls/intrinsics/diagnostics.rs
@@ -7,6 +7,50 @@ pub fn add_intrinsics(impls: &mut Implementations) {
     };
 }
 
+// ------ Benchmark ------------------------------------------------------------
+fn benchmark_help() -> HelpMessage {
+    HelpMessage {
+        desc: "time the expression evaluation\npipes <input> to <expr>".into(),
+        params: vec![HelpParameter::Required("expr".into())],
+        examples: vec![
+            HelpExample {
+                desc: "time loading in a file",
+                code: "benchmark { open file.csv }",
+            },
+            HelpExample {
+                desc: "time filtering a table",
+                code: "\\ file.csv | benchmark filter = col 1",
+            },
+        ],
+        ..HelpMessage::new("benchmark")
+    }
+}
+
+fn benchmark_intrinsic(mut blk: Block) -> Result<Step> {
+    let expr = blk.next_arg(None)?;
+    blk.eval_o(move |val, cx| {
+        let start = Instant::now();
+        expr.resolve(|| val, &cx)?;
+        cx.done_o(benchmark_table(start, Instant::now()))
+    })
+}
+
+fn benchmark_table(start: Instant, end: Instant) -> Table {
+    let d = end - start;
+    let (ms, us) = (d.subsec_millis(), d.subsec_micros());
+    let us = us - (ms * 1000);
+    ::table::Table::from(vec![
+        vec![
+            o("duration"),
+            o("seconds"),
+            o("milliseconds"),
+            o("microseconds"),
+        ],
+        vec![o(format!("{:?}", d)), n(d.as_secs()), n(ms), n(us)],
+    ])
+    .into()
+}
+
 // ------ Typify ---------------------------------------------------------------
 fn typify_help() -> HelpMessage {
     HelpMessage {
